сделал одну ячейку на моковой фотке
