+++ ImageFeedIos/ImageList/ImageListCell.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+final class ImagesListCell: UITableViewCell {
+    static let reuseIdentifier = "ImagesListCell"
+}
