--- ImageFeedIos/View/ImageListCell.swift
@@ -1,3 +1,0 @@
-import UIKit
-
-
